+++ cli/src/lib/storage/mapping-type.ts
