+++ cli/src/lib/storage/memory-snapshot.ts
