+++ cli/src/lib/storage/types/array-type.ts
