+++ cli/src/lib/storage/types/fixed-array-type.ts
