+++ cli/src/lib/storage/types/struct-type.ts
