+++ cli/src/lib/storage/values/empty-value.ts
