+++ cli/src/lib/storage/values/mapping-value.ts
