+++ cli/src/lib/storage/values/struct-value.ts
