+++ cli/test/memory-map.spec.ts
@@ -1,18 +1,18 @@
 import { describe, expect, it } from "vitest";
-import { StorageChanges } from "../src/lib/storage-changes/storage-changes";
+import { StorageChanges } from "../src/lib/storage/storage-changes";
 import fs from "node:fs/promises";
 import path from "node:path";
 import { memoryDiffParser } from "../src/schema/rpc";
 import { bytesToHex, type Hex, hexToBigInt } from "viem";
-import { AddressType } from "../src/lib/storage-changes/types/address-type";
-import { StructType } from "../src/lib/storage-changes/types/struct-type";
-import { BigNumberType } from "../src/lib/storage-changes/types/big-number-type";
-import { Property } from "../src/lib/storage-changes/property";
-import { BooleanType } from "../src/lib/storage-changes/types/boolean-type";
+import { AddressType } from "../src/lib/storage/types/address-type";
+import { StructType } from "../src/lib/storage/types/struct-type";
+import { BigNumberType } from "../src/lib/storage/types/big-number-type";
+import { Property } from "../src/lib/storage/property";
+import { BooleanType } from "../src/lib/storage/types/boolean-type";
 import type { MemoryReport } from "../src/lib/reports/memory-report";
-import { type PropertyChange } from "../src/lib/storage-changes/property-change";
-import type { MemoryValue } from "../src/lib/storage-changes/values/memory-value";
-import type { ValueField } from "../src/lib/storage-changes/values/struct-value";
+import { type PropertyChange } from "../src/lib/storage/property-change";
+import type { MemoryValue } from "../src/lib/storage/values/memory-value";
+import type { ValueField } from "../src/lib/storage/values/struct-value";
 import { Option } from "nochoices";
 
 class TestReport implements MemoryReport<string> {
